# docs/swagger.yaml
basePath: /api
definitions:
  client.AddRestaurantItemRequest:
    properties:
      description:
        type: string
      image:
        type: string
      name:
        type: string
      notes:
        type: string
      price:
        type: string
      type:
        $ref: '#/definitions/enum.ItemType'
    required:
    - name
    - price
    - type
    type: object
  client.AddRestaurantTableRequest:
    properties:
      label:
        type: string
      notes:
        type: string
    type: object
  client.ChangePasswordRequest:
    properties:
      confirmNewPassword:
        type: string
      newPassword:
        type: string
    required:
    - confirmNewPassword
    - newPassword
    type: object
  client.CreateRestaurantOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/client.RestaurantItemOrderRequest'
        type: array
      notes:
        type: string
    required:
    - items
    type: object
  client.CreateRestaurantRequest:
    properties:
      contactEmail:
        type: string
      contactPhone:
        type: string
      description:
        type: string
      image:
        items:
          type: string
        type: array
      location:
        type: string
      name:
        type: string
      notes:
        type: string
      postalCode:
        type: string
      type:
        type: string
    required:
    - location
    - name
    - notes
    - postalCode
    - type
    type: object
  client.CreateSubPackageRequest:
    properties:
      description:
        type: string
      durationMonth:
        type: integer
      image:
        type: string
      name:
        type: string
      price:
        type: string
    required:
    - durationMonth
    - name
    - price
    type: object
  client.OrderResponse:
    properties:
      amount:
        type: number
      id:
        type: integer
      orderDate:
        type: string
      restaurantId:
        type: integer
      status:
        $ref: '#/definitions/enum.OrderStatus'
      tenantId:
        type: integer
    type: object
  client.PayOrderWithCashRequest:
    properties:
      notes:
        type: string
      orderId:
        type: integer
    required:
    - orderId
    type: object
  client.PaySubPackageRequest:
    properties:
      restaurantId:
        type: integer
      subPackageId:
        type: integer
    required:
    - restaurantId
    - subPackageId
    type: object
  client.RestaurantItemOrderRequest:
    properties:
      item:
        type: string
      quantity:
        type: integer
    required:
    - item
    - quantity
    type: object
  client.SignInRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  client.SignInResponse:
    properties:
      accessToken:
        type: string
      lastLogin:
        type: string
      refreshToken:
        type: string
    type: object
  client.TenantInfoRequest:
    properties:
      address:
        type: string
      description:
        type: string
      identity:
        type: string
      phone:
        type: string
      postalCode:
        type: string
      profileImage:
        type: string
      type:
        $ref: '#/definitions/enum.TenantType'
    required:
    - address
    - identity
    - phone
    - postalCode
    - type
    type: object
  client.TenantResponse:
    properties:
      address:
        type: string
      birthdate:
        type: string
      code:
        type: string
      description:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/enum.Gender'
      identity:
        type: string
      lastName:
        type: string
      notes:
        type: string
      phone:
        type: string
      postalCode:
        type: string
      profileImage:
        type: string
      status:
        $ref: '#/definitions/enum.UserStatus'
      type:
        $ref: '#/definitions/enum.TenantType'
      userId:
        type: integer
    type: object
  client.TenantSignUpRequest:
    properties:
      birthdate:
        type: string
      confirmPassword:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/enum.Gender'
      lastName:
        type: string
      password:
        type: string
    required:
    - birthdate
    - confirmPassword
    - email
    - firstName
    - gender
    - lastName
    - password
    type: object
  client.UpdateProfileRequest:
    properties:
      address:
        type: string
      birthdate:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/enum.Gender'
      identity:
        type: string
      lastName:
        type: string
      phone:
        type: string
      postalCode:
        type: string
      profileImage:
        type: string
    type: object
  client.UserResponse:
    properties:
      address:
        type: string
      birthdate:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/enum.Gender'
      id:
        type: integer
      identity:
        type: string
      lastName:
        type: string
      phone:
        type: string
      postalCode:
        type: string
      profileImage:
        type: string
      role:
        $ref: '#/definitions/enum.Role'
      status:
        $ref: '#/definitions/enum.UserStatus'
    type: object
  client.VerifyChangePasswordRequest:
    properties:
      verifiedCode:
        type: string
    required:
    - verifiedCode
    type: object
  enum.Gender:
    enum:
    - MALE
    - FEMALE
    type: string
    x-enum-varnames:
    - GENDER_MALE
    - GENDER_FEMALE
  enum.ItemType:
    enum:
    - FOOD
    - BEVERAGE
    type: string
    x-enum-varnames:
    - ITEM_FOOD
    - ITEM_BEVERAGE
  enum.OrderStatus:
    enum:
    - PENDING
    - COMPLETED
    - CANCELED
    type: string
    x-enum-varnames:
    - ORDER_PENDING
    - ORDER_COMPLETED
    - ORDER_CANCELED
  enum.Role:
    enum:
    - ADMIN
    - MANAGER
    - STAFF
    - TENANT
    type: string
    x-enum-varnames:
    - ROLE_ADMIN
    - ROLE_MANAGER
    - ROLE_STAFF
    - ROLE_TENANT
  enum.TenantType:
    enum:
    - PERSONAL
    - BUSINESS
    - FOOD
    - BEVERAGE
    - FB
    type: string
    x-enum-varnames:
    - TENANT_PERSONAL
    - TENANT_BUSINESS
    - RESTAURANT_FOOD
    - RESTAURANT_BEVERAGE
    - RESTAURANT_FOOD_AND_BEVERAGE
  enum.UserStatus:
    enum:
    - ACTIVE
    - INACTIVE
    - PENDING
    - LOCKED
    - DELETED
    type: string
    x-enum-varnames:
    - USER_ACTIVE
    - USER_INACTIVE
    - USER_PENDING
    - USER_LOCKED
    - USER_DELETED
host: localhost:8080
info:
  contact: {}
  description: API Documentation for FB Chain Management API - Developed by Ducking
    Team
  title: FB Chain Management API
  version: "1.0"
paths:
  /admin/subscription:
    post:
      consumes:
      - application/json
      parameters:
      - description: CreateSubPackage body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.CreateSubPackageRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create Subscription Package API
  /auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: ChangePassword body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.ChangePasswordRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Change password API
  /auth/change-password/verify:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
      security:
      - BearerAuth: []
      summary: Get change password OTP API
    post:
      consumes:
      - application/json
      parameters:
      - description: VerifyChangePassword body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.VerifyChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
      security:
      - BearerAuth: []
      summary: Verify change password OTP API
  /auth/signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: SignIn body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.SignInResponse'
        "400":
          description: Bad Request
          schema: {}
      summary: Sign in API
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: TenantSignUp body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.TenantSignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.UserResponse'
        "400":
          description: Bad Request
          schema: {}
      summary: Tenant sign up API
  /payment/cash:
    post:
      consumes:
      - application/json
      parameters:
      - description: PayOrderWithCash body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.PayOrderWithCashRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
      security:
      - BearerAuth: []
      summary: Pay order with cash API
  /payment/online/{method}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Online method
        enum:
        - VNPAY
        in: path
        name: method
        required: true
        type: string
      - description: Order ID
        in: query
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Pay order with online payment API
  /profile:
    patch:
      consumes:
      - application/json
      parameters:
      - description: UpdateProfile body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.UserResponse'
        "400":
          description: Bad Request
          schema: {}
      security:
      - BearerAuth: []
      summary: Change Profile API
  /restaurant:
    post:
      consumes:
      - application/json
      parameters:
      - description: CreateRestaurant body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.CreateRestaurantRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create restaurant API
  /restaurant/{restaurantId}/item:
    post:
      consumes:
      - application/json
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantId
        required: true
        type: string
      - description: AddRestaurantItem body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.AddRestaurantItemRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Add New Restaurant Item API
  /restaurant/{restaurantId}/table:
    post:
      consumes:
      - application/json
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantId
        required: true
        type: string
      - description: AddRestaurantTable body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.AddRestaurantTableRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Add New Restaurant Table API
  /restaurant/{restaurantId}/table/{tableId}/qrCode:
    get:
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Restaurant ID
        in: path
        name: restaurantId
        required: true
        type: string
      responses: {}
      security:
      - BearerAuth: []
      summary: Get Restaurant Table QR Code
  /table/{tableId}/order:
    post:
      consumes:
      - application/json
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: CreateRestaruantOrder body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.CreateRestaurantOrderRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create Restaurant Order API
  /table/{tableId}/session/end:
    get:
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      responses: {}
      security:
      - BearerAuth: []
      summary: End Table Ordering Session API
  /table/{tableId}/session/start:
    get:
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      responses: {}
      security:
      - BearerAuth: []
      summary: Start Table Ordering Session API
  /tenant/order:
    post:
      consumes:
      - application/json
      parameters:
      - description: PaySubPackage body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.PaySubPackageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/client.OrderResponse'
        "400":
          description: Bad Request
          schema: {}
      security:
      - BearerAuth: []
      summary: Pay subscription package API
  /tenant/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: TenantInfo body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.TenantInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.TenantResponse'
        "400":
          description: Bad Request
          schema: {}
      security:
      - BearerAuth: []
      summary: Complete Tenant Info API
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"